Perform CC check within click callback
This is a workaround for double fired checkout form submit events.
Before this workaround the CC check was performed within the first form
submit event. After the CC check a second form submit was fired to
actually finish the form submit. The commit changes this behaviour that
the CC check will be performed within a click event bound to the submit
button. The CC check callback will then trigger the submit event to submit
the form.